refactored force-velocity chart
