+++ vitaly_frontend/src/App.tsx
@@ -113,93 +113,6 @@ const App: React.FC = () => {
     return () => clearInterval(intervalId);
   }, []);
 
-  // Chart init
-  useEffect(() => {
-    // Force-Velocity Chart
-    const forceVelocityChartDom = document.getElementById("forceVelocityChart");
-    if (forceVelocityChartDom) {
-      const forceVelocityChart = echarts.init(forceVelocityChartDom);
-      const option = apiData
-        ? {
-            grid: { top: 20, right: 20, bottom: 40, left: 50 },
-            xAxis: {
-              type: "value",
-              name: "Velocity (m/s)",
-              nameLocation: "middle",
-              nameGap: 25,
-              min: 0,
-              max: 3,
-            },
-            yAxis: {
-              type: "value",
-              name: "Force (N)",
-              nameLocation: "middle",
-              nameGap: 35,
-              min: 0,
-              max: 1000,
-            },
-            series: [
-              {
-                name: "Current Session",
-                type: "line",
-                data: [[apiData.velocity, apiData.force]], // 单个数据点
-                smooth: true,
-                symbolSize: 8,
-                lineStyle: { color: "#3B82F6", width: 3 },
-                itemStyle: { color: "#3B82F6" },
-              },
-            ],
-          }
-        : {
-            grid: { top: 20, right: 20, bottom: 40, left: 50 },
-            xAxis: {
-              type: "value",
-              name: "Velocity (m/s)",
-              nameLocation: "middle",
-              nameGap: 25,
-              min: 0,
-              max: 3,
-            },
-            yAxis: {
-              type: "value",
-              name: "Force (N)",
-              nameLocation: "middle",
-              nameGap: 35,
-              min: 0,
-              max: 1000,
-            },
-            series: [
-              {
-                // Previous session curve
-                type: "line",
-                data: previousSessionData.force.map((force, index) => [
-                  previousSessionData.velocity[index],
-                  force,
-                ]),
-                smooth: true,
-                symbolSize: 6,
-                lineStyle: { color: "#E5E7EB", width: 2 },
-                itemStyle: { color: "#E5E7EB" },
-              },
-              {
-                // Current session curve
-                type: "line",
-                data: velocityData.force.map((force, index) => [
-                  velocityData.velocity[index],
-                  force,
-                ]),
-                smooth: true,
-                symbolSize: 8,
-                lineStyle: { color: "#3B82F6", width: 3 },
-                itemStyle: { color: "#3B82F6" },
-              },
-            ],
-          };
-    
-      forceVelocityChart.setOption(option);
-    }
-  }, [velocityData, previousSessionData, apiData]);
-
   // correlation chart
   useEffect(() => {
     const correlationChartDom = document.getElementById("correlationChart");
@@ -323,7 +236,11 @@ const App: React.FC = () => {
         <EmgChart emgData={emgData} />
 
         {/* Muscle Force & Power Analysis */}
-        <ForceAnalysis apiData={apiData} />
+        <ForceAnalysis 
+          apiData={apiData} 
+          velocityData={velocityData}
+          previousSessionData={previousSessionData}
+        />
 
         {/* Muscle Activation & Efficiency */}
         <MuscleActivation mvcValue={mvcValue} correlationChartId="correlationChart" />
